chore: seperating isb and pipeline pagination
(cherry picked from commit ec9f5f9b4a589406d96e121a3079bb65794e2fe8)
